Tune the feedforward lead in SERVO_TUNE_DYNAMICS (TERMS=LEAD)
The ferr fits keep saying the FF lands late (positive onset bias and a
13-sigma jerk coefficient through every bench round) and no
mass/viscous/coulomb value can express a timing error - and the bench
A/B confirmed it (lead 0 is worse than 1). The lead was a
bringup-frozen integer cycle count, but the endpoint applies it as
rings[s].peek_vel_acc(apply_time + ff_lead_ns[s]) - a peek at an
arbitrary future nanosecond - so it was never physically quantized.

- ec-rt: SetFfLead{slot,lead_ns}/SetFfLeadResponse (0x008E/0x008F)
  runtime command writing ctx.ff_lead_ns[slot]; invalid slot fails
  with -309 like SetDriveLimits; stub acks it.
- --ff-lead-cycles and [motor] ff_lead_cycles accept fractional
  cycles (f64 end to end; bringup rounds to ns).
- engine.set_ff_lead(handle, slot, lead_ns) pyo3 method, mirroring
  set_diff_damper.
- SERVO_TUNE_DYNAMICS: LEAD joins TERMS (now in the default list) as
  one shared node-global continuous line search in seconds - scored on
  the mean of both modes' rms, first direction from the summed onset
  bias (positive = late = probe up), floored at zero with a half-cycle
  floor step, half-cycle zero-floor step. Requires one shared
  ff_lead_cycles across the tuned servos (fails loudly otherwise).
  The winner streams to every slot and stays live until RESTART; the
  converged report and manifest carry lead_us/lead_cycles so the value
  can be persisted via ff_lead_cycles. Aborts restore the configured
  lead along with the baseline model.

+++ rust/motion-engine/src/bridge/state.rs
@@ -144,7 +144,7 @@ pub(crate) struct EthercatDrive {
     pub(crate) max_torque_tenth_pct: Option<u16>,
     pub(crate) velocity_ff: bool,
     pub(crate) ff_max_torque: f64,
-    pub(crate) ff_lead_cycles: u32,
+    pub(crate) ff_lead_cycles: f64,
     pub(crate) invert_direction: bool,
     pub(crate) dynamics_profile: Option<String>,
 }
